Try to increase PlayerScreen update accuracy

+++ Assets/JANOARG.Client/Scripts/Behaviors/Player/PlayerScreenPause.cs
@@ -23,7 +23,7 @@ namespace JANOARG.Client.Behaviors.Player
 
         public Image RetryFlash;
 
-        public float PauseTime = -10;
+        public double PauseTime = -10;
 
         [Space] public Coroutine CurrentAnimation;
